# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_KhalidMohomud_ecomApi_internal_dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      slug:
        type: string
    type: object
...
  github_com_KhalidMohomud_ecomApi_internal_dto.CreateCategoryRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      name:
        maxLength: 150
        minLength: 2
        type: string
      parent_id:
        type: string
      slug:
        maxLength: 170
        type: string
    required:
    - name
    type: object
...
  github_com_KhalidMohomud_ecomApi_internal_dto.PaginatedResponse-github_com_KhalidMohomud_ecomApi_internal_dto_CategoryResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_dto.CategoryResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_KhalidMohomud_ecomApi_internal_dto.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 150
        minLength: 2
        type: string
      parent_id:
        type: string
      slug:
        maxLength: 170
        type: string
    required:
    - name
    type: object
...
  github_com_KhalidMohomud_ecomApi_internal_utils.ErrorResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_KhalidMohomud_ecomApi_internal_utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: alizakifarah@gmail.com
    name: API Support
  description: Production-ready E-commerce REST API built with Go, Gin, GORM, and
    PostgreSQL.
  license:
    name: MIT
  title: E-commerce API
  version: "1.0"
paths:
...
  /categories:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_dto.PaginatedResponse-github_com_KhalidMohomud_ecomApi_internal_dto_CategoryResponse'
              type: object
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Admin only. If slug is omitted, it's generated from name.
      parameters:
      - description: Category payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_dto.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.ErrorResponse'
        "409":
          description: slug already in use
          schema:
            $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - categories
  /categories/{id}:
    delete:
      description: Admin only. Soft-deletes the category; any child categories are
        promoted to top-level.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - categories
    get:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_dto.CategoryResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.ErrorResponse'
      summary: Get a category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Admin only. Full replace — every field is overwritten with the
        request body.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_dto.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
  /categories/slug/{slug}:
    get:
      parameters:
      - description: Category slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_dto.CategoryResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_KhalidMohomud_ecomApi_internal_utils.ErrorResponse'
      summary: Get a category by slug
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token, e.g.
      "Bearer eyJhbGciOi...".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"